requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-openai
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
pypdf
docx2txt
google-search-results
